Backport Code, rename package to th-compat
The new name is somewhat more appropriate than `th-quote`, as this
package covers more than just `Quote` now.
