+++ tests/Language/Haskell/TH/Syntax/CompatSpec.hs
@@ -1,14 +1,15 @@
+{-# LANGUAGE CPP #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE TemplateHaskell #-}
 
--- | Regression tests for "Language.Haskell.TH.Syntax.Quote".
-module Language.Haskell.TH.Syntax.QuoteSpec (main, spec) where
+-- | Regression tests for "Language.Haskell.TH.Syntax.Compat".
+module Language.Haskell.TH.Syntax.CompatSpec (main, spec) where
 
 import Control.Exception (evaluate)
 import Control.Monad.State
 
 import Language.Haskell.TH.Syntax hiding (newName)
-import Language.Haskell.TH.Syntax.Quote
+import Language.Haskell.TH.Syntax.Compat
 
 import Prelude ()
 import Prelude.Compat
@@ -32,6 +33,12 @@ spec = parallel $ do
     it "errors on Quasi-ful expressions" $
       evaluate (runPureQ (unsafeQToQuote (qReport True "Explosion in 3... 2... 1...")))
         `shouldThrow` errorCall "`unsafeQToQuote` does not support code that uses qReport"
+
+#if MIN_VERSION_template_haskell(2,9,0)
+  describe "IsCode" $
+    it "manipulates typed TH expressions in a backwards-compatible way" $
+      $$(fromCode (toCode [|| "abc" ||])) `shouldBe` "abc"
+#endif
 
 newtype PureQ a = MkPureQ (State Uniq a)
   deriving (Functor, Applicative, Monad, MonadState Uniq)
